Look at producer handling of encoding too.
Producer API already includes a method accepting a Writer, so no need
to add anything except a bit of JavaDoc on the method that takes File,
explaining how the encoding is chosen, since someone using the API needs
to understand that.

Used the opportunity to tweak the TapProducer so that any inability to
encode the output in the chosen encoding will result in exceptions,
instead of Java's old default silently-drop-or-corrupt ... when you call
newEncoder() explicitly, you get one that throws exceptions, while all the
other Java API methods to select an encoding give the old lossy behavior.